chore: export and move generateOpenSshKeyPair to utils/
